Add a deploy stage (OpenJDK 11 only)
The deploy runs only in OpenJDK 11.
The artifact version is taken from the tag name.
Targets are rebuilt, but not tested.

We call `mvn package`, which takes care of installation etc.

We use the `jdk` key instead of calling `jdk_switcher`, because
jdk_switcher seems to be broken.

This has been reported a few times, with no resolution:
- https://travis-ci.community/t/jdk-switcher-command-not-found/3364
- https://travis-ci.community/t/rocketmq-travis-env-issue-command-not-found/3496